requirements.txt
boto3==1.35.8
botocore==1.35.8
envyaml==1.10.211231
prometheus_client==0.17.0
python_dateutil==2.8.2
